+++ jannovar-core/src/main/java/de/charite/compbio/jannovar/annotation/Annotation.java
@@ -137,7 +137,7 @@ public final class Annotation implements Comparable<Annotation> {
 	public Annotation(TranscriptModel transcript, GenomeChange change, Collection<VariantEffect> varTypes,
 			AnnotationLocation annoLoc, String ntHGVSDescription, String aaHGVSDescription,
 			Collection<AnnotationMessage> messages) {
-		this.change = change;
+		this.change = change.withStrand('+'); // enforce forward strand
 		if (varTypes == null)
 			this.effects = ImmutableSortedSet.<VariantEffect> of();
 		else